remocao 'acabada'

+++ src/remotion.c
@@ -1,10 +1,13 @@
 #include <stdlib.h>
 #include <stdio.h>
+
+#include <string.h> // strcmp
+#include <stdbool.h>
 #include <utils.h>
 #include <search.h>
 
 int binary_search(int *vector, int begin, int end, int val) {
-	if (end < begin) return -1;
+	if (end < begin) return INVALID;
 	int aux = (begin + end)/2;
 
 	if (vector[aux] == val) return aux;
@@ -12,38 +15,164 @@ int binary_search(int *vector, int begin, int end, int val) {
 	return binary_search(vector, aux + 1, end, val);
 }
 
+
 void print_menu_remove() {
 	printf("Please, give the ticket number to be removed\n");
 	printf(">> ");
 }
 
-void logical_remove(FILE *fp, int pos, int top) {
-	int counter = 0, invalid = -1;
-	char aux = 'a';
 
-	// Go to 'pos'
-	fseek(fp, pos, SEEK_SET);
+/* Returns the next element in the linked list.
+If the element is not invalid, also return by nextSize its size. */
+int next_element(FILE *fp, int byteOffset, int *nextSize) {
+	int next;
+	
+	fseek(fp, byteOffset, SEEK_SET); // Go to the current position 
+	fseek(fp, 2*sizeof(int), SEEK_CUR); // skip over invalid and reg size
+	
+	fread(&next, sizeof(int), 1, fp); // gets the next element in the linked list
+	
+	fseek(fp, next, SEEK_SET); // Go to the next element
+	fseek(fp, sizeof(int), SEEK_CUR); // skip over invalid 
+	fread(nextSize, sizeof(int), 1, fp); // reads the size of the next element
+	
+	return next;
+}
 
-	// Go to the end of the record to count it's size
-	while (!feof(fp) && aux != '#') {
-		fread(&aux, sizeof(char), 1, fp);
-		if (!feof(fp)) counter++;
-	}
 
-	// Go back to pos
-	fseek(fp, pos, SEEK_SET);
+
+
+void mark_reg_invalid(FILE *fp, int byteOffset, int nextElement) {
+	int regSize, invalid = INVALID;
+
+	// Gets the size of the register
+	regSize = get_register_size(fp, byteOffset);
+
+	// Go back to byteOffset
+	fseek(fp, byteOffset, SEEK_SET);
 
 	// Mark the record as invalid
 	fwrite(&invalid, sizeof(int), 1, fp);
 
 	// Mark the record's size
-	fwrite(&counter, sizeof(int), 1, fp);
-
+	fwrite(&regSize, sizeof(int), 1, fp);
+	
 	// Mark the next in the list
-	fwrite(&top, sizeof(int), 1, fp);
+	fwrite(&nextElement, sizeof(int), 1, fp);
 }
 
-void remove_index(FILE *fp, int ticket, char *file_name) {
+
+int logical_remove_best_fit_search(FILE *fp, int byteOffset, t_list *list) {
+	int pos, next, nextSize;
+	int regSize;
+	
+	regSize = get_register_size(fp, byteOffset);
+	
+	// Go to the given position
+	fseek(fp, byteOffset, SEEK_SET);
+	
+	// pos starts at the beginning of the list
+	pos = list->head;
+	
+	while (pos != INVALID ) {
+		next = next_element(fp, pos, &nextSize);
+		
+		// if the next element has a smaller (or equal) regsize, go to it
+		if (next != INVALID && nextSize <= regSize)
+			pos = next;
+		
+		// we found as far we can go
+		else
+			break;
+	}
+	
+	// if we didnt find a position in the list, at it to the end of the data file.
+	return get_file_size(fp);
+}
+
+
+
+int logical_remove_worst_fit_search(FILE *fp, int byteOffset, t_list *list) {
+	int pos, next, nextSize;
+	int regSize;
+	
+	regSize = get_register_size(fp, byteOffset);
+	
+	// Go to the given position
+	fseek(fp, byteOffset, SEEK_SET);
+	
+	// pos starts at the beginning of the list
+	pos = list->head;
+	
+	while (pos != INVALID ) {
+		next = next_element(fp, pos, &nextSize);
+		
+		// if the next element has a bigger (or equal) regsize, go to it
+		if (next != INVALID && nextSize >= regSize)
+			pos = next;
+		
+		// we found as far we can go
+		else
+			break;
+	}
+	
+	// if we didnt find a position in the list, at it to the end of the data file.
+	return get_file_size(fp);
+}
+
+
+
+void logical_remove_best_and_worst(FILE *fp, int byteOffset, t_list *list, char *type) {
+	int listOffset, regSize, next, invalid = INVALID;
+	
+	regSize = get_register_size(fp, byteOffset);
+	
+	// If the current linked list is empty
+	if (list->head == INVALID) { 
+		// Add it as a root
+		mark_reg_invalid(fp, byteOffset, list->head);
+		list->head = byteOffset;
+		return;
+	}
+
+	// Find the position in the linked list that we need to add the new element
+	if ( strcmp(type, "best") == 0)
+		listOffset = logical_remove_best_fit_search(fp, byteOffset, list);
+	else
+		listOffset = logical_remove_worst_fit_search(fp, byteOffset, list);
+		
+	// If its not the new list root
+	if (listOffset != -1) {
+		// Goes to the position we will be adding a new element after
+		fseek(fp, listOffset, SEEK_SET);
+	
+		fseek(fp, 2*sizeof(int), SEEK_CUR); // skips the invalid and sizeIndicator
+		fread(&next, sizeof(int), 1, fp);	// gets the position of the next element
+		
+		// pos->next = new element
+		fseek(fp, -sizeof(int), SEEK_CUR); 	// we go back to the rewrite the "->next" element
+		fwrite(&byteOffset, sizeof(int), 1, fp);
+		
+		// new element -> next = next
+		mark_reg_invalid(fp, byteOffset, next);			
+	}
+	
+	// It is the list root
+	else {
+		mark_reg_invalid(fp, byteOffset, list->head);
+		list->head = byteOffset;
+	}
+}
+
+
+
+void logical_remove_first(FILE *fp, int byteOffset, t_list *list) {
+	// Adds the removed register to the beginning of the linked list
+	mark_reg_invalid(fp, byteOffset, list->head);	
+}
+
+
+bool remove_index(FILE *fp, int ticket, char *file_name) {
 	int i, fileSize, limitTicketBO, count, index;
 	int *tickets, *byteOffsets;
 
@@ -66,6 +195,12 @@ void remove_index(FILE *fp, int ticket, char *file_name) {
 
 	// Search for the index of the ticket
 	index = binary_search(tickets, 0, count-1, ticket);
+	
+	// The ticket was not found
+	if (index == INVALID) {
+		printf("That ticket was not found, aborting \n");
+		return false;
+	}
 
 	// Shift the vector
 	for (i = index + 1; i < count; i++) {
@@ -75,42 +210,65 @@ void remove_index(FILE *fp, int ticket, char *file_name) {
 
 	// Clear the file
 	fclose(fp);
-	fp = fopen(file_name, "w"); // pode dar bug porque o ponteiro pode ser outro que nao estava no fp antes, precisa testar
+	fp = fopen(file_name, "w"); // TODO pode dar bug porque o ponteiro pode ser outro que nao estava no fp antes, precisa testar
 
-	// Write the content back to the file
+	// Write the content back to the index file
 	fwrite(tickets, count-1, sizeof(int), fp);
 	fwrite(byteOffsets, count-1, sizeof(int), fp);
 
 	// Free the vector
 	free(tickets);
 	free(byteOffsets);
+	
+	return true;
 }
 
-void remove_record(t_files *files, t_list *list) {
-	int ticket, byteOffset;
 
+void remove_record(t_files *files, t_list *lists) {
+	int ticket, byteOffset;
+	char *temp;
+	bool found;
+
+	// TODO: acho que esse scanf da ruim qnd digitado strings e outras coisas aleatorias (B)
 	while (1) {
 		print_menu_remove();
-		scanf("%d", &ticket);
+		
+		// reads the user input
+	    temp = read_line(stdin, FIELD_DELIM, LINE_END, VARIABLE_FIELD);
+	    ticket = atoi(temp);
+	    free(temp);
+	    
 		if (ticket > 0) break;
-		printf("Ticket must be a integer greater then zero\n");
+		printf("Ticket must be a integer greater then zero!!\n\n\n");
 	}
 
 	// Best fit
-	// byteOffset = busca_no_arquivo_de_indice_best
-	// logial_remove(files->outputBest, byteOffset, list->best);
-	remove_index(files->indexBest, ticket, "best.idx");
-	// list->best = byteOffset;
+	found = search_primary_index(files->indexBest, ticket, &byteOffset);
+	if ( found ) {
+		remove_index(files->indexBest, ticket, "best.idx");
+		logical_remove_best_and_worst(files->outputBest, byteOffset, &(lists[BEST]), "best");
+	}
+	else
+		printf("Ticket was not found in best.idx\n");
+
 
 	// Worst fit
-	// byteOffset = busca_no_arquivo_de_indice_worst
-	// logial_remove(files->outputWorst, byteOffset, list->worst);
-	remove_index(files->indexWorst, ticket, "worst.idx");
-	// list->worst = byteOffset;
+	found = search_primary_index(files->indexWorst, ticket, &byteOffset);
+	if (found) {
+		remove_index(files->indexWorst, ticket, "worst.idx");
+		logical_remove_best_and_worst(files->outputWorst, byteOffset, &(lists[WORST]), "worst");
+	}
+	else
+		printf("Ticket was not found in worst.idx\n");
+
 
 	// First fit
-	// byteOffset = busca_no_arquivo_de_indice_first
-	// logial_remove(files->outputFirst, byteOffset, list->first);
-	remove_index(files->indexFirst, ticket, "first.idx");
-	// list->first = byteOffset;
+	found = search_primary_index(files->indexFirst, ticket, &byteOffset);
+	if (found) {
+		logical_remove_first(files->outputFirst, byteOffset, &(lists[FIRST]));
+		remove_index(files->indexFirst, ticket, "first.idx");
+	}
+	else
+		printf("Ticket was not found in first.idx\n");
+		
 }
